+++ project/.env.local
@@ -1,0 +1,2 @@
+PORT=9999
+DISABLE_ESLINT_PLUGIN=true
