+++ project/src/components/app/page-modification.ts
@@ -1,0 +1,6 @@
+enum PAGE_MODIFICATION {
+  MAIN = 'page--gray page--main',
+  LOGIN = 'page--gray page--login',
+}
+
+export default PAGE_MODIFICATION;
