+++ project/src/components/app/routes-paths.ts
@@ -1,0 +1,9 @@
+enum ROUTES_PATHS {
+  MAIN = '/',
+  LOGIN = '/login',
+  FAVORITES = '/favorites',
+  ROOM = '/offer/:id',
+  ANYTHING = '*',
+}
+
+export default ROUTES_PATHS;
