+++ project/src/components/footer/footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function Footer() {
+  return (
+    <footer className="footer container">
+      <a className="footer__logo-link" href="main.html">
+        <img className="footer__logo" src="./img/logo.svg" alt="6 cities logo" width="64" height="33" />
+      </a>
+    </footer>
+  );
+}
+
+export default Footer;
