+++ project/src/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import App from './components/app/app';
 
@@ -7,6 +8,8 @@ const ARENDA_OFFER_DATA = [0, 1, 2, 3, 4];
 
 ReactDOM.render(
   <React.StrictMode>
-    <App arendaOfferData={ARENDA_OFFER_DATA} />
+    <BrowserRouter>
+      <App arendaOfferData={ARENDA_OFFER_DATA} />
+    </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root'));
